Compile with overflow checks in release
I tend to develop my solution using the test example, running in debug, and then run the real input with release. Unfortunately, quite often the real input is larger in a way that causes overflows in number types that were big enough for the example. Since these wrap in release, I might miss it and submit an incorrect solution.
The runtime cost of running with overflow checks in release wasn't that high, so now I'm running with some guardrails.